requirements.txt
quart
pydantic>=2.0
discord.py>=2.3
aiohttp
pyyaml
